+++ services/server/app/api/users/models.py
@@ -1,10 +1,7 @@
 from app import bcrypt
 from app import db
 from app.api.users.admin import UserAdminView
-from flask import current_app
 
-import datetime
-import jwt
 import os
 
 
@@ -23,35 +20,6 @@ class User(db.Model):
         self.password = bcrypt.generate_password_hash(
             password, current_app.config.get("BCRYPT_LOG_ROUNDS")
         ).decode()
-
-    def encode_token(self, user_id, token_type):
-        config = current_app.config
-        time_to_live = (
-            config.get("ACCESS_TOKEN_EXPIRATION")
-            if token_type == "access"
-            else config.get("REFRESH_TOKEN_EXPIRATION")
-        )
-        payload = {
-            "exp": datetime.datetime.utcnow()
-            + datetime.timedelta(seconds=time_to_live),
-            "iat": datetime.datetime.utcnow(),
-            "sub": user_id,
-        }
-        return jwt.encode(
-            payload=payload,
-            key=config.get("SECRET_KEY"),
-            algorithm=config.get("JWT_ENCODE_ALGORITHM"),
-        )
-
-    @staticmethod
-    def decode_token(token):
-        config = current_app.config
-        payload = jwt.decode(
-            jwt=token,
-            key=config.get("SECRET_KEY"),
-            algorithms=[config.get("JWT_ENCODE_ALGORITHM")],
-        )
-        return payload["sub"]
 
 
 if os.getenv("FLASK_ENV") != "production":
